REFACTOR rename argumentSuggester to argumentSuggestor
